requirements.txt
streamlit==1.28.2
google-genai
opencv-python-headless
numpy
Pillow
streamlit-drawable-canvas